// src/api/openapi/models/curriculum-summary-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Zeroone API v1
 * v1
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 커리큘럼 요약 항목
 */
export interface CurriculumSummaryDto {
    /**
     * 미션 ID
     */
    'missionId'?: number;
    /**
     * 미션 제목
     */
    'title'?: string;
    /**
     * 주차
     */
    'weekNum'?: number;
}